Move PointIndex to mir_dataflow.
